refactor(tasks): separate tasks and messages
